Update payment gateways types and messages
