+++ src/@next/components/organisms/AdyenPaymentGateway/AdyenPaymentGateway.tsx
@@ -1,10 +1,16 @@
 import { CompleteCheckout_checkoutComplete_order } from "@saleor/sdk/lib/mutations/gqlTypes/CompleteCheckout";
 import React, { useEffect, useRef, useState } from "react";
-import { defineMessages, IntlShape, useIntl } from "react-intl";
+import { useIntl } from "react-intl";
 
 import { ErrorMessage } from "@components/atoms";
 import { paymentErrorMessages } from "@temp/intl";
-import { IFormError, IPaymentGatewayConfig } from "@types";
+import {
+  IFormError,
+  IPaymentGatewayConfig,
+  IPaymentSubmitResult,
+} from "@types";
+
+import { adyenErrorMessages } from "./intl";
 
 export const adyenNotNegativeConfirmationStatusCodes = [
   "Authorised",
@@ -14,55 +20,6 @@ export const adyenNotNegativeConfirmationStatusCodes = [
   "Received",
   "PresentToShopper",
 ];
-
-const messageDescription = "Adyen payment gateway error";
-
-export const adyenErrorMessages = defineMessages({
-  unknownPayment: {
-    defaultMessage: "Unknown payment submission error occured.",
-    description: messageDescription,
-  },
-  invalidPaymentSubmission: {
-    defaultMessage: "Invalid payment submission.",
-    description: messageDescription,
-  },
-});
-
-export const adyenConfirmationErrorMessages = defineMessages({
-  error: {
-    defaultMessage: "Error processing payment occured.",
-    description: messageDescription,
-  },
-  refused: {
-    defaultMessage:
-      "The payment was refused. Try the payment again using a different payment method or card.",
-    description: messageDescription,
-  },
-  cancelled: {
-    defaultMessage: "Payment was cancelled.",
-    description: messageDescription,
-  },
-  general: {
-    defaultMessage: "Payment confirmation went wrong.",
-    description: messageDescription,
-  },
-});
-
-export function translateAdyenConfirmationError(
-  status: string,
-  intl: IntlShape
-): string {
-  switch (status) {
-    case "Error":
-      return intl.formatMessage(adyenConfirmationErrorMessages.error);
-    case "Refused":
-      return intl.formatMessage(adyenConfirmationErrorMessages.refused);
-    case "Cancelled":
-      return intl.formatMessage(adyenConfirmationErrorMessages.cancelled);
-    default:
-      return intl.formatMessage(adyenConfirmationErrorMessages.general);
-  }
-}
 
 interface IResourceConfig {
   src: string;
@@ -105,12 +62,12 @@ export interface IProps {
   /**
    * Method to call on gateway payment submission.
    */
-  submitPayment: (data: object) => Promise<any>;
+  submitPayment: (data: object) => Promise<IPaymentSubmitResult>;
   /**
    * Method called after succesful gateway payment submission. This is the case when no confirmation is needed.
    */
   submitPaymentSuccess: (
-    order?: CompleteCheckout_checkoutComplete_order
+    order?: CompleteCheckout_checkoutComplete_order | null
   ) => void;
   /**
    * Errors returned by the payment gateway.
