+++ src/@next/components/organisms/AdyenPaymentGateway/index.ts
@@ -1,1 +1,2 @@
 export * from "./AdyenPaymentGateway";
+export * from "./intl";
