+++ src/@next/components/organisms/PaymentGatewaysList/types.ts
@@ -1,6 +1,11 @@
 import { CompleteCheckout_checkoutComplete_order } from "@saleor/sdk/lib/mutations/gqlTypes/CompleteCheckout";
 
-import { ICardData, IFormError, IPaymentGateway } from "@types";
+import {
+  ICardData,
+  IFormError,
+  IPaymentGateway,
+  IPaymentSubmitResult,
+} from "@types";
 
 export interface IProps {
   /**
@@ -39,9 +44,9 @@ export interface IProps {
     token?: string,
     cardData?: ICardData
   ) => void;
-  submitPayment: (data?: object) => Promise<any>;
+  submitPayment: (data?: object) => Promise<IPaymentSubmitResult>;
   submitPaymentSuccess: (
-    order?: CompleteCheckout_checkoutComplete_order
+    order?: CompleteCheckout_checkoutComplete_order | null
   ) => void;
   /**
    * Method called when gateway error occured.
