+++ src/@next/components/organisms/StripePaymentGateway/stories.tsx
@@ -9,7 +9,7 @@ const config = [
   { field: "api_key", value: "pk_test_6pRNASCoBOKtIshFeQd4XMUh" },
 ];
 const processPayment = action("processPayment");
-const submitPayment = async () => action("submitPayment");
+const submitPayment = async () => Promise.resolve({});
 const submitPaymentSuccess = action("submitPaymentSuccess");
 const onError = action("onError");
