+++ src/@next/components/organisms/StripePaymentGateway/types.ts
@@ -1,6 +1,11 @@
 import { CompleteCheckout_checkoutComplete_order } from "@saleor/sdk/lib/mutations/gqlTypes/CompleteCheckout";
 
-import { ICardData, IFormError, IPaymentGatewayConfig } from "@types";
+import {
+  ICardData,
+  IFormError,
+  IPaymentGatewayConfig,
+  IPaymentSubmitResult,
+} from "@types";
 
 export interface IProps {
   /**
@@ -26,12 +31,12 @@ export interface IProps {
   /**
    * Method to call on gateway payment submission.
    */
-  submitPayment: (data?: object) => Promise<any>;
+  submitPayment: () => Promise<IPaymentSubmitResult>;
   /**
    * Method called after succesful gateway payment submission. This is the case when no confirmation is needed.
    */
   submitPaymentSuccess: (
-    order?: CompleteCheckout_checkoutComplete_order
+    order?: CompleteCheckout_checkoutComplete_order | null
   ) => void;
   /**
    * Method called when gateway error occured.
