+++ src/@next/types/IPaymentGateway.ts
@@ -1,3 +1,7 @@
+import { CompleteCheckout_checkoutComplete_order } from "@saleor/sdk/lib/mutations/gqlTypes/CompleteCheckout";
+
+import { IFormError } from "@types";
+
 export interface IPaymentGatewayConfig {
   /**
    * Gateway config key.
@@ -22,4 +26,11 @@ export interface IPaymentGateway {
    * Payment gateway client configuration.
    */
   config: IPaymentGatewayConfig[];
+}
+
+export interface IPaymentSubmitResult {
+  confirmationData?: string | null;
+  confirmationNeeded?: boolean;
+  order?: CompleteCheckout_checkoutComplete_order | null;
+  errors?: IFormError[];
 }
